worked on API integration for Courses

+++ src/services/admin/courses.js
@@ -3,7 +3,7 @@ import api from "@services/api";
 // Create a new course
 export const createCourse = async (courseData) => {
   try {
-    const response = await api.post("admin/add-courses", courseData);
+    const response = await api.post("/admin/courses", courseData);
     return response.data;
   } catch (error) {
     throw error.response?.data || error.message;
@@ -13,7 +13,7 @@ export const createCourse = async (courseData) => {
 // Update an existing course
 export const updateCourse = async (id, courseData) => {
   try {
-    const response = await api.put(`/courses/${id}`, courseData);
+    const response = await api.put(`/admin/courses/${id}`, courseData);
     return response.data;
   } catch (error) {
     throw error.response?.data || error.message;
@@ -23,7 +23,7 @@ export const updateCourse = async (id, courseData) => {
 // Delete a course
 export const deleteCourse = async (id) => {
   try {
-    const response = await api.delete(`/courses/${id}`);
+    const response = await api.delete(`/admin/courses/${id}`);
     return response.data;
   } catch (error) {
     throw error.response?.data || error.message;
@@ -37,5 +37,23 @@ export const getCourseById = async (id) => {
     return response.data;
   } catch (error) {
     throw error.response?.data || error.message;
+  }
+};
+
+export const getCourses = async () => {
+  try {
+    const response = await api.get("/user/courses");
+    return response;
+  } catch (error) {
+    throw error.response.data;
+  }
+};
+
+export const getCourseCategories = async () => {
+  try {
+    const response = await api.get("/user/categories");
+    return response;
+  } catch (error) {
+    throw error.response.data;
   }
 };
